requirements.txt
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
Django==2.0.4
djangorestframework==3.8.2
get==1.0.2
httpie==0.9.9
idna==2.6
libmagic==1.0
post==1.0.1
public==1.0.1
Pygments==2.2.0
python-dateutil==2.7.2
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2018.4
query-string==1.0.1
request==1.0.1
requests==2.18.4
six==1.11.0
urllib3==1.22
